requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
matplotlib>=3.3.0
seaborn>=0.11.0
pytest>=6.0.0
python-tk>=0.1.0
